% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sn_capabilities}
\alias{sn_capabilities}
\title{Return negotiated ShennongDB capabilities}
\usage{
sn_capabilities(connection = sn_connection(), refresh = FALSE)
}
\arguments{
\item{connection}{A ShennongDB connection.}

\item{refresh}{Whether to renegotiate capabilities.}
}
\value{
The negotiated capabilities.
}
\description{
Return negotiated ShennongDB capabilities
}
