% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sn_connection}
\alias{sn_connection}
\title{Return the default ShennongDB connection}
\usage{
sn_connection()
}
\value{
The current default ShennongDB connection.
}
\description{
Return the default ShennongDB connection
}
