% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sn_disconnect}
\alias{sn_disconnect}
\title{Remove a session ShennongDB connection}
\usage{
sn_disconnect(connection = sn_connection())
}
\arguments{
\item{connection}{A ShennongDB connection.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Remove a session ShennongDB connection
}
