% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sn_ping}
\alias{sn_ping}
\title{Ping ShennongDB}
\usage{
sn_ping(connection = sn_connection())
}
\arguments{
\item{connection}{A ShennongDB connection.}
}
\value{
The server version response.
}
\description{
Ping ShennongDB
}
