% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sn_set_connection}
\alias{sn_set_connection}
\title{Set the default ShennongDB connection}
\usage{
sn_set_connection(connection)
}
\arguments{
\item{connection}{A ShennongDB connection.}
}
\value{
\code{connection}, invisibly.
}
\description{
Set the default ShennongDB connection
}
