% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sn_slice_head}
\alias{sn_slice_head}
\title{Limit a lazy ShennongData query}
\usage{
sn_slice_head(x, n = 6L, ...)
}
\arguments{
\item{x}{A \link{ShennongData} handle.}

\item{n}{Maximum number of observations requested per feature.}

\item{...}{Reserved for compatible methods.}
}
\value{
A lazy \link{ShennongData} handle.
}
\description{
Limit a lazy ShennongData query
}
